Make the utility reusable.
